Implemented FeatureFlagsApiService and modified unit test accordingly.
